Using rm -rf on a variable without ensuring it is set can be dangerous. It is safer to use the ${VAR:?} syntax to ensure the script exits if the variable is unset or null, preventing accidental deletion of unintended directories if the script is modified in the future.

+++ create-dmg.sh
@@ -15,7 +15,7 @@ if [ ! -d "$ARTIFACTS_DIR" ]; then
 fi
 
 STAGING_DIR="dmg_staging"
-rm -rf "$STAGING_DIR"
+rm -rf "${STAGING_DIR:?}"
 mkdir -p "$STAGING_DIR"
 
 # Copy bundles
